fix: gitignore and include change
the build process failed because
github-pages 223 | Error:  Could not locate the included file 'notes_graph.html' in any of ["/github/workspace/_includes"]. Ensure it exists in one of those directories and is not a symlink as those are not allowed in safe mode.
replacing it with  m.html in content.html to test whether the builder considers any liquied tag even if enclosed in comments.